Separate examples in different files
