+++ example/encode_actions.dart
@@ -1,0 +1,42 @@
+import 'package:dart_esr/dart_esr.dart';
+
+main(List<String> args) => actionsExample();
+
+Future<void> actionsExample() async {
+  var auth = <Authorization>[ESRConstants.PlaceholderAuth];
+
+  var data = <String, String>{'name': 'data'};
+
+  var action = Action()
+    ..account = 'eosnpingpong'
+    ..name = 'ping'
+    ..authorization = auth
+    ..data = data;
+
+  var action2 = Action()
+    ..account = 'eosnpingpong'
+    ..name = 'ping'
+    ..authorization = auth
+    ..data = data;
+
+  var action3 = Action()
+    ..account = 'eosnpingpong'
+    ..name = 'ping'
+    ..authorization = auth
+    ..data = data;
+
+  var args = SigningRequestCreateArguments(
+    actions: [action, action2, action3],
+    chainId: 'e70aaab8997e1dfce58fbfac80cbbb8fecec7b99cf982a9444273cbc64c41473',
+  );
+  SigningRequestEncodingOptions options = defaultSigningRequestEncodingOptions(
+      nodeUrl: 'https://jungle.greymass.com');
+
+  try {
+    var request = await SigningRequestManager.create(args, options: options);
+    var uri = request.encode();
+    print('actions\n' + uri);
+  } catch (e) {
+    print(e.toString());
+  }
+}
