+++ example/example.dart
@@ -1,0 +1,13 @@
+import 'deserialize_signing_request.dart';
+import 'encode_action.dart';
+import 'encode_actions.dart';
+import 'encode_identity.dart';
+import 'encode_transaction.dart';
+
+main(List<String> args) async {
+  transactionExample();
+  actionExample();
+  actionsExample();
+  identityExample();
+  desezializeExample();
+}
